Fix windows warnings (#71)
* Fix windows warnings

(cherry picked from commit e1616665f239f49668ae23e9b437cd4538d2db9d)

+++ topic_tools/include/topic_tools/drop_node.hpp
@@ -35,8 +35,8 @@ public:
 private:
   void process_message(std::shared_ptr<rclcpp::SerializedMessage> msg) override;
 
-  int x_;
-  int y_;
+  int64_t x_;
+  int64_t y_;
   int count_{0};
 };
 }  // namespace topic_tools
